use gate to cotroller

+++ app/Http/Controllers/Admins/ClassroomController.php
@@ -19,6 +19,7 @@ class ClassroomController extends Controller
 
     public function index(Request $request)
     {
+        $this->authorize('view-classroom');
         $data = $this->classroomRepository->search($request->only(['name', 'subject', 'faculty']));
         return view('backends.classrooms.index')->with([
             'classrooms' => $data['classrooms'],
@@ -30,6 +31,7 @@ class ClassroomController extends Controller
 
     public function create()
     {
+        $this->authorize('create-classroom');
         $data=$this->classroomRepository->create();
         return view('backends.classrooms.create')->with(['faculties'=>$data['faculty'], 'subjects'=>$data['subject']]);
     }
@@ -52,6 +54,7 @@ class ClassroomController extends Controller
 
     public function edit($id)
     {
+        $this->authorize('edit-classroom');
         $data = $this->classroomRepository->edit($id);
         return view('backends.classrooms.edit')->with(['faculties'=>$data['faculty'], 'subjects'=>$data['subject'],'classroom'=>$data['classroom']]);
 
@@ -66,6 +69,7 @@ class ClassroomController extends Controller
 
     public function destroy($id)
     {
+        $this->authorize('destroy-classroom');
         return response()->json([
             'status' => 204,
             'message' => $this->classroomRepository->destroy($id),
